Address Razvan's comments
